Consuming less memory

+++ test/neighborhood.js
@@ -466,16 +466,16 @@ describe('Neighborhood Indices', function() {
       assert.strictEqual(index.level, 1);
       assert.deepEqual(index.counts.slice(0, index.C), new Uint8Array([1, 1]));
       assert.deepEqual(index.neighborhood.slice(0, index.E), new Uint8Array([1, 0]));
-      assert.deepEqual(index.weights.slice(0, index.E), new Float64Array([1, 1]));
+      assert.deepEqual(index.weights.slice(0, index.E), new Uint8Array([1, 1]));
       assert.deepEqual(index.starts.slice(0, index.C + 1), new Uint8Array([0, 1, 2]));
       assert.deepEqual(index.belongings.slice(0, index.C), new Uint8Array([0, 1]));
-      assert.deepEqual(index.totalWeights.slice(0, index.C), new Float64Array([3, 9]));
+      assert.deepEqual(index.totalWeights.slice(0, index.C), new Uint8Array([3, 9]));
       // assert.deepEqual(index.internalWeights.slice(0, index.C), new Float64Array([2, 8]));
 
       // Once more
       index.move(0, 1, 1);
 
-      assert.deepEqual(index.totalWeights.slice(0, index.C), new Float64Array([0, 12]));
+      assert.deepEqual(index.totalWeights.slice(0, index.C), new Uint8Array([0, 12]));
       // assert.deepEqual(index.internalWeights.slice(0, index.C), new Float64Array([0, 12]));
       assert.strictEqual(index.U, 1);
       assert.deepEqual(index.unused.slice(0, index.U), new Uint8Array([0]));
@@ -821,7 +821,7 @@ describe('Neighborhood Indices', function() {
 
       index.zoomOut();
 
-      assert.deepEqual(index.totalWeights.slice(0, index.C), new Float64Array([6, 6]));
+      assert.deepEqual(index.totalWeights.slice(0, index.C), new Uint8Array([6, 6]));
       // assert.deepEqual(index.internalWeights.slice(0, index.C), new Float64Array([2, 2]));
     });
 
@@ -873,7 +873,7 @@ describe('Neighborhood Indices', function() {
 
       assert.strictEqual(index.E, 2);
       // assert.deepEqual(index.internalWeights.slice(0, index.C), new Float64Array([4, 2]));
-      assert.deepEqual(index.totalWeights.slice(0, index.C), new Float64Array([8, 6]));
+      assert.deepEqual(index.totalWeights.slice(0, index.C), new Uint8Array([8, 6]));
 
       var UQ = index.modularity();
